Rakesh, Vineela | MOBN-1369 | Refactor test folder structure
